backend/requirements.txt
cat > requirements.txt << 'EOF'
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
pillow==10.1.0
openai==1.54.0
requests==2.31.0
EOF
